Спецификация API для Task Engine web + исправление timestamps на int

+++ sandbox/extensions/task_engine/models.py
@@ -27,8 +27,8 @@ class TaskStatusResult(BaseModel):
     attempt_no: int
     partial_result: str | None = None
     error: str | None = None
-    created_at: float = 0.0
-    updated_at: float = 0.0
+    created_at: int = 0
+    updated_at: int = 0
 
 
 class ActiveTasksResult(BaseModel):
@@ -103,8 +103,8 @@ class TaskRecord:
     schedule_at: float | None
     leased_by: str | None
     lease_exp: float | None
-    created_at: float
-    updated_at: float
+    created_at: int
+    updated_at: int
     after_task_id: str | None = None
     chain_id: str | None = None
     chain_order: int | None = None
@@ -125,4 +125,4 @@ class StepRecord:
     tokens_used: int | None = None
     duration_ms: int | None = None
     error_code: str | None = None
-    created_at: float = 0.0
+    created_at: int = 0
